+++ test/fixtures/components/a_component.ex
@@ -2,6 +2,7 @@ defmodule AComponent do
   use Phoenix.Component
 
   def a_component(assigns) do
-    ~H"<span>a component: <%= @label %></span>"
+    assigns = assign_new(assigns, :index, fn -> 42 end)
+    ~H"<span data-index={@index}>a component: <%= @label %></span>"
   end
 end