+++ test/fixtures/components/b_component.ex
@@ -2,6 +2,13 @@ defmodule BComponent do
   use Phoenix.LiveComponent
 
   def render(assigns) do
-    ~H"<span>b component: <%= @label %></span>"
+    ~H"""
+    <span>
+      b component: <%= @label %>
+      <%= if assigns[:block] do %>
+        <%= @block %>
+      <% end %>
+    </span>
+    """
   end
 end