+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## v0.2.0 (unreleased yet)
+
+- added new tab to browse your component sources
+- added work-in-progress component documentation source
+
 ## v0.1.0 (2022-07-26)
 
 Initial release.